requirements.txt
# API e Banco de Dados
requests==2.31.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0

# Apache Airflow
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-http==4.5.1